Resolve conflicts by adopting master's fuller JsonRepository/main.cpp
(Read/Update/Delete already merged there) while keeping this branch's
UTF-8 console/source-encoding fix so Korean text renders correctly.

+++ Cluade_CRUD/Record.cpp
@@ -5,7 +5,7 @@ using ClaudeJson::JsonValue;
 JsonValue ToJson(const Record& record)
 {
     JsonValue json(JsonValue::Object{});
-    json["id"] = JsonValue(record.id);
+    json["id"] = JsonValue(static_cast<double>(record.id));
     json["name"] = JsonValue(record.name);
     json["value"] = JsonValue(record.value);
     return json;
